Implement front controller
