+++ api/user/urls.py
@@ -1,5 +1,5 @@
 from django.conf.urls import url
-from user.views import (
+from views import (
     AddNewUser,
     DeleteUser)
