+++ api/user_token/urls.py
@@ -1,7 +1,7 @@
 from django.conf.urls import url
 from django.conf import settings
 from django.conf.urls.static import static
-from user_token.views import ValidateToken
+from views import ValidateToken
 from rest_framework.authtoken import views as rest_framework_views
 
 urlpatterns = [
