+++ api/views.py
@@ -1,3 +1,7 @@
+from rest_framework.authtoken.models import Token
+from django.shortcuts import get_object_or_404
+from django.http import JsonResponse
+from rest_framework.views import APIView
 from user.models import User
 from rest_framework import generics
 from rest_framework.authentication import TokenAuthentication
@@ -5,6 +9,24 @@ from user.serializers import UserSerializer
 from rest_framework.permissions import (
     AllowAny,
     IsAuthenticated)
+
+
+class ValidateToken(APIView):
+    permission_classes = [AllowAny]
+
+    def get(self, request):
+        return JsonResponse({'status': False,
+                             'message': "Método get não permitido"})
+
+    def post(self, request):
+        data = request.data
+        token = data['token']
+        obj = get_object_or_404(Token, key=token)
+        if obj is not None:
+            return JsonResponse({'status': True})
+
+        message = "Token not found !"
+        return JsonResponse({'status': False, 'message': message})
 
 
 class AddNewUser(generics.CreateAPIView):
